+++ newton/examples/uipc/example_uipc_hello_world.py
@@ -17,7 +17,7 @@ import newton.examples
 
 
 class Example:
-    def __init__(self, viewer:newton.viewer.ViewerBase, args):
+    def __init__(self, viewer: newton.viewer.ViewerBase, args):
         self.fps = 60
         self.frame_dt = 1.0 / self.fps
         self.sim_time = 0.0
@@ -53,13 +53,17 @@ class Example:
             pitch=0.0,
             yaw=-180.0,
         )
-        self.viewer._paused = True 
+        self.viewer._paused = True
 
     def simulate(self):
         for _ in range(self.sim_substeps):
             self.state_0.clear_forces()
             self.solver.step(
-                self.state_0, self.state_1, self.control, self.contacts, self.sim_dt,
+                self.state_0,
+                self.state_1,
+                self.control,
+                self.contacts,
+                self.sim_dt,
             )
             self.state_0, self.state_1 = self.state_1, self.state_0